Bug fixes, confirmation dialgs
Fixed bugs, added adaptive grid to admin panel, moved order managing page to admin panel

+++ src/pages/adminPanel/categoriesPanel/categoriesList.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="row pl-3 pr-3">
+    <div class="row">
         <div class="col-12 h2">
             <i class="fas fa-clipboard-list"></i>
             Список категорій
@@ -8,9 +8,15 @@
             <div class="row list">
                 <div class="col-12 list-item" v-for="category in categories">
                     <div class="row">
-                        <div class="col-4">{{category.categoryName}}</div>
-                        <div class="col-4">{{category.categoryUrl}}</div>
-                        <div class="col-4">
+                        <div class="col-12 col-lg-5">
+                            <span class="font-weight-bold d-lg-none">Назва: </span>
+                            <span>{{category.categoryName}}</span>
+                        </div>
+                        <div class="col-12 col-lg-5">
+                            <span class="font-weight-bold d-lg-none">URL: </span>
+                            <span>{{category.categoryUrl}}</span>
+                        </div>
+                        <div class="col">
                             <div class="row justify-content-end">
                                 <router-link :to="'/admin-panel/categories/update/' + category.categoryUrl" tag="div" class="col-auto">
                                     <i class="fas fa-edit"></i>
@@ -42,14 +48,23 @@
                 deleteCategoryById: 'deleteCategory'
             }),
             deleteCategory(id) {
-                this.deleteCategoryById(id)
-                    .then(() => {
-                        this.showSuccessAlert("Категорію видалено")
-                        this.loadCategories();
-                    })
-                    .catch((error) => {
-                        this.showErrorMessage("Невдалось видалити категорію", error)
-                    })
+                this.showConfirmDialog(
+                    "Ви впевнені що хочете видалити категорію?",
+                    "Якщо у категорії присутні товари, її не можливо видалити.",
+                    "Спробувати видалити"
+                ).then((result) => {
+                    if (result.value) {
+                        this.deleteCategoryById(id)
+                            .then(() => {
+                                this.showSuccessAlert("Категорію видалено");
+                                this.loadCategories();
+                            })
+                            .catch((error) => {
+                                this.showErrorMessage("Невдалось видалити категорію", error)
+                            })
+                    }
+                })
+
             },
             showSuccessAlert(title) {
                 this.$swal.fire({
@@ -65,6 +80,16 @@
                     title,
                     text
                 })
+            },
+            showConfirmDialog(title, text, confirmText) {
+                return this.$swal.fire({
+                    title,
+                    text,
+                    type: 'warning',
+                    showCancelButton: true,
+                    confirmButtonText: confirmText,
+                    cancelButtonText: 'Скасувати'
+                });
             }
         },
         created() {
